Major Feature: Pin Favorite Non-owned Folders Previously, when visiting the folder of another user, only that user's folders are shown.
Added feature such that the logged-in user's folders are also shown (side-by-side with the non-owned folders).
The tab is tinted, a tooltip is added to specify ownership, and non-owned folders are protected.
Each non-owned folder has a checkbox. If checked, the folder will appear any time user is logged in, along with owned folders.

Set "GlowscriptDemos/Examples" as default pinned folder. Was unable to test without web deployment, but should work like any user.